+++ app/domain/project_materials/__init__.py
@@ -1,0 +1,8 @@
+from .entities import ProjectMaterial
+from .errors import ProjectMaterialNotFoundError, ProjectMaterialValidationError
+
+__all__ = [
+    "ProjectMaterial",
+    "ProjectMaterialNotFoundError",
+    "ProjectMaterialValidationError",
+]